requirements.txt
alabaster==0.7.16
anyio==4.3.0
anytree==2.12.1
attrs==23.2.0
autopage==0.5.2
Babel==2.15.0
beautifulsoup4==4.12.3
bullet==2.2.0
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cliff==4.6.0
cmd2==2.4.3
colorama==0.4.6
configobj==5.0.8
configparser==7.0.0
distlib==0.3.8
dnspython==2.0.0
docutils==0.21.2
filelock==3.14.0
gnupg==2.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
imagesize==1.4.1
ipwhois==1.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
mccabe==0.7.0
motor==3.4.0
packaging==24.0
pbr==6.0.0
pexpect==4.9.0
platformdirs==4.2.1
pluggy==1.5.0
prettytable==3.10.0
property-manager==3.0
psutil==5.9.8
ptyprocess==0.7.0
pyflakes==3.2.0
Pygments==2.18.0
pymongo==4.7.2
pyperclip==1.8.2
pyproject-api==1.6.1
pyreadline3==3.4.1
python-dotenv==1.0.1
python-secrets==23.4.2
python-telegram-bot==21.1.1
PyYAML==6.0.1
requests==2.31.0
setuptools-scm==8.1.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stevedore==5.2.0
tox==4.15.0
urllib3==2.2.1
verboselogs==1.7
virtualenv==20.26.2
wcwidth==0.2.13
xkcdpass==1.19.9
yamlreader==3.0.4
